Merging changes to the pipelines for No t2w (#2)
* Pipelines for sessions without T2w images, for BIDS data.

First version.  I am making sure the scripts do not crash if there is no T2w image, and I am also using "fsl_anat" to to the BiasCorrection.

* Fixes Atlas Registration for sessions without T2w images.

The Atlas Registration in fsl_anat is not that great (maybe because the brain extraction is not good), so run it separatedly.

* Fixed FreeSurferPipeline for sessions without T2w images.

It skips the steps that require T2w images. The pipeline runs, but I am not sure if there will be some missing output files down the pipeline.

* Fixed PostFreeSurferPipeline for sessions without T2w images.

It skips the steps that require T2w images. The pipeline runs, but I am not sure if there will be some missing output files down the pipeline.

* Fixed bug in call to AtlasRegistration...

Fixed bug in call to AtlasRegistrationToMNI152_FLIRTandFNIRT in the case of no T2w images.

* Fixes typo in T2w2T1Reg

Minor typo in logging messages.

* Fixed PreFreeSurferPipeline for sessions without T2w images.

It fixes the BiasField computation by using the smoothing defined in the pipeline, and putting the bias field image where the rest of the pipelines expect it.

* Fixed PreFreeSurferPipeline for sessions without T2w images.

It fixes the images output by the bias field correction for the sessions with no T2w.

* Updating pipelines for sessions without T2w images.

Fixed some bugs.

* Updating pipelines for sessions without T2w images.

Fixed some bugs I forgot to include in the previous commit.

* Updating PostFreeSurfer pipeline for sessions without T2w images.

Fixed a bug.

* Updates Dockerfile ahead of the merge w v3.4

* Updates Dockerfile ahead of the merge w v3.4

Just a comment in the Dockerfile